requirements.txt
scipy
scikit-learn
pytorch
numpy
torchmetrics